+++ R/MultiLevelIPF-package.R
@@ -8,8 +8,8 @@
 #' \tabular{ll}{
 #'   Package: \tab MultiLevelIPF\cr
 #'   Type: \tab Package\cr
-#'   Version: \tab 0.0-11.2\cr
-#'   Date: \tab 2014-11-15\cr
+#'   Version: \tab 0.0-11.3\cr
+#'   Date: \tab 2014-11-20\cr
 #'   License: \tab GPL v3\cr
 #' } ~~
 #' An overview of how to use the package, including the most